+++ saltedge-ios-swift/Classes/API/Models/Parameters/SEAccountParams.swift
@@ -24,13 +24,16 @@
 import Foundation
 
 public struct SEAccountParams: URLEncodable, ParametersEncodable {
+    let perPage: String
     let fromId: Int?
     
-    public init(fromId: Int?) {
+    public init(perPage: String = "1000", fromId: Int?) {
+        self.perPage = perPage
         self.fromId = fromId
     }
 
     enum CodingKeys: String, CodingKey {
+        case perPage = "per_page"
         case fromId = "from_id"
     }
 }